Divide up MainPage.tsx into separate components
