+++ src/components/App.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Fragment } from 'react';
-import Home from './Home';
+import Main from './Main/main';
 
 export interface IAppProps {
 }
@@ -9,9 +9,7 @@ export default class App extends React.Component<IAppProps> {
   public render() {
     return (
       <Fragment>
-        <h1>Hello, World!</h1>
-        <Home/>
-
+     <Main/>
       </Fragment>
     );
   }
